+++ software/lamp-os/src/behaviors/dmx.cpp
@@ -1,5 +1,7 @@
 #include "./dmx.hpp"
 
+#include <cstdint>
+
 #include "../util/color.hpp"
 
 namespace lamp {
@@ -12,8 +14,7 @@ void DmxBehavior::control() {};
 
 void DmxBehavior::setColor(Color inColor) { currentColor = inColor; };
 
-void DmxBehavior::setLastArtnetFrameTimeMs(
-    unsigned long inLastArtnetFrameTimeMs) {
+void DmxBehavior::setLastArtnetFrameTimeMs(uint32_t inLastArtnetFrameTimeMs) {
   lastArtnetFrameTimeMs = inLastArtnetFrameTimeMs;
 };
 }  // namespace lamp
