+++ software/lamp-os/src/behaviors/dmx.hpp
@@ -1,6 +1,8 @@
 #ifndef LAMP_BEHAVIORS_DMX_H
 #define LAMP_BEHAVIORS_DMX_H
 
+#include <cstdint>
+
 #include "../components/network/wifi.hpp"
 #include "../core/animated_behavior.hpp"
 #include "../util/color.hpp"
@@ -14,7 +16,7 @@ class DmxBehavior : public AnimatedBehavior {
 
  public:
   Color currentColor = Color();
-  unsigned long lastArtnetFrameTimeMs = 0;
+  uint32_t lastArtnetFrameTimeMs = 0;
 
   void draw();
 
@@ -22,7 +24,7 @@ class DmxBehavior : public AnimatedBehavior {
 
   void setColor(Color inColor);
 
-  void setLastArtnetFrameTimeMs(unsigned long inLastArtnetFrameTimeMs);
+  void setLastArtnetFrameTimeMs(uint32_t inLastArtnetFrameTimeMs);
 };
 }  // namespace lamp
 #endif
