+++ software/lamp-os/src/util/levels.cpp
@@ -1,14 +1,16 @@
 #include "./levels.hpp"
 
+#include <cstdint>
+
 namespace lamp {
-int darken(int value, int percentage) {
-  int p = 100 - ((percentage * 100) / 100);
+uint8_t darken(uint8_t value, uint8_t percentage) {
+  uint8_t p = 100 - ((percentage * 100) / 100);
 
   return (value * p) / 100;
 };
 
-int calculateBrightnessLevel(int value, int percentage) {
-  int p = ((percentage * 100) / 100);
+uint8_t calculateBrightnessLevel(uint8_t value, uint8_t percentage) {
+  uint8_t p = ((percentage * 100) / 100);
 
   return (value * p) / 100;
 };
